+++ shells/cmd.go
@@ -1,8 +1,15 @@
 package shells
 
 import (
-	"errors"
+	"bufio"
+	"bytes"
+	"fmt"
 	"github.com/ayufan/gitlab-ci-multi-runner/common"
+	"github.com/ayufan/gitlab-ci-multi-runner/helpers"
+	"io"
+	"path/filepath"
+	"runtime"
+	"strings"
 )
 
 type CmdShell struct {
@@ -12,8 +19,105 @@ func (c *CmdShell) GetName() string {
 	return "cmd"
 }
 
-func (c *CmdShell) GenerateScript(build *common.Build) (*common.ShellScript, error) {
-	return nil, errors.New("not yet supported")
+func (b *CmdShell) writeCommand(w io.Writer, format string, args ...interface{}) {
+	io.WriteString(w, fmt.Sprintf(format, args...)+"\r\n")
+}
+
+func (b *CmdShell) writeCommandChecked(w io.Writer, format string, args ...interface{}) {
+	b.writeCommand(w, format, args...)
+	b.writeCommand(w, "%s", "IF %errorlevel% NEQ 0 exit /b %errorlevel%")
+}
+
+func (b *CmdShell) writeCloneCmd(w io.Writer, build *common.Build) {
+	b.writeCommand(w, "echo Clonning repository...")
+	b.writeCommandChecked(w, "md %s", filepath.FromSlash(build.BuildsDir))
+	b.writeCommandChecked(w, "cd %s", filepath.FromSlash(build.BuildsDir))
+	b.writeCommandChecked(w, "rd /s /q %s", filepath.FromSlash(build.ProjectDir()))
+	b.writeCommandChecked(w, "git clone %s %s", build.RepoURL, filepath.FromSlash(build.ProjectDir()))
+	b.writeCommandChecked(w, "cd %s", filepath.FromSlash(build.ProjectDir()))
+}
+
+func (b *CmdShell) writeFetchCmd(w io.Writer, build *common.Build) {
+	b.writeCommand(w, "IF EXIST \"%s\\%s\\.git\" (", filepath.FromSlash(build.BuildsDir), filepath.FromSlash(build.ProjectDir()))
+	b.writeCommand(w, "echo Fetching changes...")
+	b.writeCommandChecked(w, "cd %s\\%s", filepath.FromSlash(build.BuildsDir), filepath.FromSlash(build.ProjectDir()))
+	b.writeCommandChecked(w, "git clean -fdx")
+	b.writeCommandChecked(w, "git reset --hard > NUL")
+	b.writeCommandChecked(w, "git remote set-url origin %s", build.RepoURL)
+	b.writeCommandChecked(w, "git fetch origin")
+	b.writeCommand(w, ") ELSE (")
+	b.writeCloneCmd(w, build)
+	b.writeCommand(w, ")")
+}
+
+func (b *CmdShell) writeCheckoutCmd(w io.Writer, build *common.Build) {
+	b.writeCommand(w, "echo Checkouting %s as %s...", build.Sha[0:8], build.RefName)
+	b.writeCommandChecked(w, "git checkout -B %s %s > NUL", build.RefName, build.Sha)
+	b.writeCommandChecked(w, "git reset --hard %s > NUL", build.Sha)
+}
+
+func (b *CmdShell) GenerateScript(build *common.Build) (*common.ShellScript, error) {
+	var buffer bytes.Buffer
+	w := bufio.NewWriter(&buffer)
+
+	b.writeCommand(w, "@echo off")
+	b.writeCommand(w, "echo.")
+	b.writeCommand(w, "setlocal enableextensions")
+
+	if len(build.Hostname) != 0 {
+		b.writeCommand(w, "echo Running on %s via %s...", "%COMPUTERNAME%", helpers.ShellEscape(build.Hostname))
+	} else {
+		b.writeCommand(w, "echo Running on %s...", "%COMPUTERNAME%")
+	}
+
+	if build.AllowGitFetch {
+		b.writeFetchCmd(w, build)
+	} else {
+		b.writeCloneCmd(w, build)
+	}
+
+	b.writeCheckoutCmd(w, build)
+
+	for _, command := range strings.Split(build.Commands, "\n") {
+		command = strings.TrimRight(command, " \t\r\n")
+		if strings.TrimSpace(command) == "" {
+			b.writeCommand(w, "echo.")
+			continue
+		}
+
+		if !build.Runner.DisableVerbose {
+			b.writeCommand(w, "echo %s", command)
+		}
+		b.writeCommandChecked(w, "%s", command)
+	}
+
+	w.Flush()
+
+	env := []string{
+		fmt.Sprintf("CI_BUILD_REF=%s", build.Sha),
+		fmt.Sprintf("CI_BUILD_BEFORE_SHA=%s", build.BeforeSha),
+		fmt.Sprintf("CI_BUILD_REF_NAME=%s", build.RefName),
+		fmt.Sprintf("CI_BUILD_ID=%d", build.ID),
+		fmt.Sprintf("CI_BUILD_REPO=%s", build.RepoURL),
+
+		fmt.Sprintf("CI_PROJECT_ID=%d", build.ProjectID),
+		fmt.Sprintf("CI_PROJECT_DIR=%s", filepath.FromSlash(build.FullProjectDir())),
+
+		"CI_SERVER=yes",
+		"CI_SERVER_NAME=GitLab CI",
+		"CI_SERVER_VERSION=",
+		"CI_SERVER_REVISION=",
+	}
+
+	script := common.ShellScript{
+		Environment: env,
+		Script:      buffer.Bytes(),
+		Command:     "cmd",
+		Arguments:   []string{"/Q", "/C"},
+		PassFile:    true,
+		Extension:   "cmd",
+	}
+	return &script, nil
 }
 
 func (b *CmdShell) IsDefault() bool {